Re-order variants to prioritize narrower types
As of https://github.com/pybind/pybind11/pull/5879, pybind11 will
convert Python `int` to C `float`/`double`. For `std::variant`, pybind11
will attempt to convert arguments in the order of the type.

So if `double` occurs earlier in the variant, then `int`/`long` will
never be produced. By putting `int`/`long` before `double`, pybind11
will attempt that conversion first and we'll continue to produce our
deprecation warning correctly.

Fixes #31495

+++ src/_backend_agg_wrapper.cpp
@@ -58,8 +58,8 @@ PyRendererAgg_draw_path(RendererAgg *self,
 static void
 PyRendererAgg_draw_text_image(RendererAgg *self,
                               py::array_t<agg::int8u, py::array::c_style | py::array::forcecast> image_obj,
-                              std::variant<double, int> vx,
-                              std::variant<double, int> vy,
+                              std::variant<int, double> vx,
+                              std::variant<int, double> vy,
                               double angle,
                               GCAgg &gc)
 {
